separate sevices in robots app
